split the canvas into multiple files
